Crate burn_dataset

Burn Dataset

Burn Dataset is a library for creating and loading datasets.

Modules

source

Sources for datasets.

transform

Transformations to be used with datasets.

Structs

DatasetIterator

Dataset iterator.

HuggingfaceDatasetLoader

Load a dataset from huggingface datasets.

InMemDataset

Dataset where all items are stored in ram.

SqliteDataset

This struct represents a dataset where all items are stored in an SQLite database. Each instance of this struct corresponds to a specific table within the SQLite database, and allows for interaction with the data stored in the table in a structured and typed manner.

SqliteDatasetStorage

The SqliteDatasetStorage struct represents a SQLite database for storing datasets. It consists of an optional name, a database file path, and a base directory for storage.

SqliteDatasetWriter

This SqliteDatasetWriter struct is a SQLite database writer dedicated to storing datasets. It retains the current writer’s state and its database connection.

Enums

ImporterError

Error type for HuggingfaceDatasetLoader.

SqliteDatasetError

Sqlite dataset error.

Traits

Dataset

The dataset trait defines a basic collection of items with a predefined size.

Type Definitions

Result

Result type for the sqlite dataset.