Struct burn_dataset::SqliteDatasetStorage

pub struct SqliteDatasetStorage { /* private fields */ }

The SqliteDatasetStorage struct represents a SQLite database for storing datasets. It consists of an optional name, a database file path, and a base directory for storage.

Implementations

impl SqliteDatasetStorage

pub fn from_name(name: &str) -> Self

Creates a new instance of SqliteDatasetStorage using a dataset name.

Arguments

name - A string slice that holds the name of the dataset.

pub fn from_file<P: AsRef<Path>>(db_file: P) -> Self

Creates a new instance of SqliteDatasetStorage using a database file path.

Arguments

db_file - A reference to the Path that represents the database file path.

pub fn with_base_dir<P: AsRef<Path>>(self, base_dir: P) -> Self

Sets the base directory for storing the dataset.

Arguments

base_dir - A string slice that represents the base directory.

pub fn exists(&self) -> bool

Checks if the database file exists in the given path.

Returns

A boolean value indicating whether the file exists or not.

pub fn db_file(&self) -> PathBuf

Fetches the database file path.

Returns

A PathBuf instance representing the file path.

pub fn base_dir(base_dir: Option<PathBuf>) -> PathBuf

Determines the base directory for storing the dataset.

Arguments

base_dir - An Option that may contain a PathBuf instance representing the base directory.

Returns

A PathBuf instance representing the base directory.

pub fn writer<I>(&self, overwrite: bool) -> Result<SqliteDatasetWriter<I>>where I: Clone + Send + Sync + Serialize + DeserializeOwned,

Provides a writer instance for the SQLite dataset.

Arguments

overwrite - A boolean indicating if the existing database file should be overwritten.

Returns

A Result which is Ok if the writer could be created, Err otherwise.

pub fn reader<I>(&self, split: &str) -> Result<SqliteDataset<I>>where I: Clone + Send + Sync + Serialize + DeserializeOwned,

Provides a reader instance for the SQLite dataset.

Arguments

split - A string slice that defines the data split for reading (e.g., “train”, “test”).

Returns

A Result which is Ok if the reader could be created, Err otherwise.